+++ CHANGELOG.md
@@ -20,3 +20,4 @@
   Adding ability to read AWS params required for provider billing.
 - adding METRICS_UTILITY_SHIP_TARGET="directory", so the billing data
   will be stored in a provided directory in daily partitions
+- adding build_report command that can build .xlsx reports locally